+++ app/Http/Controllers/GroupController.php
@@ -26,9 +26,12 @@ class GroupController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function getAll() : JsonResponse
+    public function index() : JsonResponse
     {
+        $this->authorize('index', Group::class);
+
         return new JsonResponse($this->groupRepository->findAll(), JsonResponse::HTTP_OK);
     }
 
@@ -39,9 +42,12 @@ class GroupController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      * @throws \Illuminate\Validation\ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create(Request $request) : JsonResponse
     {
+        $this->authorize('create', Group::class);
+
         $this->validate($request, [
             'number'    => ['required', 'int', 'min:3'],
             'parent_id' => ['int', 'exists:groups,id'],
@@ -63,9 +69,12 @@ class GroupController extends Controller
      * @param int $id
      *
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(int $id) : JsonResponse
     {
+        $this->authorize('show', Group::class);
+
         return new JsonResponse($this->groupRepository->find($id), JsonResponse::HTTP_OK);
     }
 
@@ -77,9 +86,12 @@ class GroupController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      * @throws \Illuminate\Validation\ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, int $id) : JsonResponse
     {
+        $this->authorize('update', Group::class);
+
         $this->validate($request, [
             'number'    => ['int', 'min:3'],
             'parent_id' => ['int', 'exists:groups,id'],
@@ -101,9 +113,12 @@ class GroupController extends Controller
      * @param int $id
      *
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(int $id) : JsonResponse
     {
+        $this->authorize('destroy', Group::class);
+
         $group = $this->groupRepository->find($id);
         $group->delete();
 
@@ -114,9 +129,12 @@ class GroupController extends Controller
      * @param int $id
      *
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function getStudents(int $id) : JsonResponse
     {
+        $this->authorize('getStudents', Group::class);
+
         return new JsonResponse(
             $this->groupRepository->getRelatedModels($id, 'students'),
             JsonResponse::HTTP_OK
@@ -127,9 +145,12 @@ class GroupController extends Controller
      * @param int $id
      *
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function getTeachers(int $id) : JsonResponse
     {
+        $this->authorize('getTeachers', Group::class);
+
         return new JsonResponse(
             $this->groupRepository->getRelatedModels($id, 'teachers'),
             JsonResponse::HTTP_OK
